refresh on card open
