+++ console/src/app/modules/onboarding/onboarding.component.scss
@@ -33,15 +33,11 @@
       margin-top: 0;
     }
 
-    .onboarding-spinner {
-      display: flex;
-      margin-bottom: 1rem;
-    }
-
     .onboarding-progress-bar-wrapper {
       display: flex;
       align-items: center;
       padding: 1rem 0;
+      min-height: 20px;
 
       .progress {
         border-radius: 50vw;
@@ -63,6 +59,7 @@
       .prog-desc {
         font-style: italic;
         font-size: 14px;
+        margin-right: 1rem;
       }
     }
